refactor(docx-reader): split document parsing into text, props, and media modules
